Add internal folder to MSCommon.  Add __init__.py, config.py, debug.py, and trace.py to internal folder.  Replace debug logging code from common.py and replace with debug method import from internal module.  Move output common code to config.py.  Move debug logging code to debug.py and trace output code to trace.py.  Rename _MSCOMMON to _MSCONFIG.  Only import trace.py module if tracing is enabled via environment variable.  Add boolean variables for debug and trace to config.py to enable/disable use without having to delete the source files.
